Replacing Ink & Splash with SwiftMarkdown2
This is a wrapper around python-markdown2, which uses Pygments under the hood. Very high quality Markdown parsing with high quality code syntax highlighting. I don't think there are Swift tools yet that match the output.

This is not ready for production yet. We're not getting the document title yet, and SwiftMarkdown2 can't be pinned to a version because PythonKit isn't versioned. That would also infect into Saga. It's not a good situation to be in.