Replaced Ink and Stencil with SwiftMarkdown (#3)
* Replacing Ink & Splash with SwiftMarkdown2

This is a wrapper around python-markdown2, which uses Pygments under the hood. Very high quality Markdown parsing with high quality code syntax highlighting. I don't think there are Swift tools yet that match the output.

This is not ready for production yet. We're not getting the document title yet, and SwiftMarkdown2 can't be pinned to a version because PythonKit isn't versioned. That would also infect into Saga. It's not a good situation to be in.

* SwiftMarkdown2 can be versioned since its dependencies are now versioned

* Using SwiftMarkdown instead of SwiftMarkdown2, due to its extension system

* Readme & changelog

* Link to PR

* Added a strikethrough test

* Added a way to supply custom SiteMetadata, which will be given to every template

* Added striptags and wordcount filters

* striptags now also removed codeblocks

* Multiple fixes

* Update to newer SwiftMarkdown, and only create a single instance instead of one for every markdown file

* The smarty extension is incredibly slow!

* swift-tools-version:5.2

* Making some things public so that we can extend readers and writers from Swift packages

* Page now also has a public initializer

* makeMetadataDecoder is public

* Minor changes

* 0.8.0 - 2021-02-07